fix(client): resolve clippy warnings (strict-but-sane gate now passes)
Make `cargo clippy --all-targets --features bin -- -D warnings` pass on the
host toolchain (1.96.0):

- Auto-applied machine-fixable lints via `cargo clippy --fix`: unused imports,
  needless borrows, derivable impls, identity maps, redundant field names,
  needless returns, unnecessary casts, etc.
- Manual straightforward fixes: rename `assetId` -> `asset_id`; build
  InitialConfiguration via struct literal (field_reassign_with_default);
  vec![] instead of Vec::new()+push; add Default for DirectoryManager;
  `if let Some` instead of is_some()+unwrap (unnecessary_unwrap); add
  `.truncate(true)` to a write-probe OpenOptions (suspicious_open_options).
- #[allow(clippy::too_many_arguments)] on three wide constructors and
  #[allow(clippy::wrong_self_convention)] on ToolMigrator::from_type
  (refactors deemed out of scope).
- Platform-conditional false positives: #[allow(unused_imports)] on
  cfg(unix) imports used only in linux-gated paths; #[allow(unreachable_code)]
  on a fn whose macOS branch returns early; #[allow(dead_code)] on the
  Windows-only service-stop constants.
- #[allow(dead_code)] (flagged for review) on genuinely-unused items:
  config::{Configuration, LoggingConfig}, permissions::ROOT_UID,
  Metric/LogShipper/VelopackUpdater unread fields, and two unused methods.
- #![allow(deprecated)] in encryption_service for aes-gcm's generic-array 0.x
  (proper fix: bump aes-gcm).
- Scope the lint gate to `--features bin` (not --all-features): the
  tool-version override features' build.rs requires release version env vars
  and is not part of the standard gate.

Windows cfg paths validate in CI (cross-compiling C deps from macOS is not
possible locally).

+++ clients/openframe-client/src/system.rs
@@ -44,8 +44,8 @@ impl SystemInfo {
         let mut disk_total = 0;
         let mut disk_used = 0;
         if let Ok(space) = sys_info::disk_info() {
-            disk_total = space.total as u64 * 1024; // Convert to bytes
-            disk_used = (space.total - space.free) as u64 * 1024;
+            disk_total = space.total * 1024; // Convert to bytes
+            disk_used = (space.total - space.free) * 1024;
         }
 
         Ok(SystemMetrics {
